+++ Notechigima/src/main/java/com/capstone/notechigima/domain/group/GroupCreateEntity.java
@@ -1,4 +1,4 @@
-package com.capstone.notechigima.domain;
+package com.capstone.notechigima.domain.group;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
