+++ Notechigima/src/main/java/com/capstone/notechigima/domain/sentence/SentenceWithWriterEntity.java
@@ -1,4 +1,4 @@
-package com.capstone.notechigima.domain;
+package com.capstone.notechigima.domain.sentence;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
