+++ Notechigima/src/main/java/com/capstone/notechigima/repository/AdviceRepository.java
@@ -1,6 +1,6 @@
 package com.capstone.notechigima.repository;
 
-import com.capstone.notechigima.domain.AdviceDetailEntity;
+import com.capstone.notechigima.domain.advice.AdviceDetailEntity;
 import org.apache.ibatis.annotations.Mapper;
 import org.springframework.stereotype.Repository;
