+++ Notechigima/src/main/resources/mapper/AdviceMapper.xml
@@ -24,10 +24,10 @@
         <result property="sentence2" column="sentence2"/>
         <result property="writerId2" column="writerId2"/>
         <result property="writerName2" column="writerName2"/>
-        <collection column="adviceId" property="comments" javaType="java.util.ArrayList" ofType="com.capstone.notechigima.domain.CommentDetailEntity" select="getCommentListById"/>
+        <collection column="adviceId" property="comments" javaType="java.util.ArrayList" ofType="com.capstone.notechigima.domain.comment.CommentDetailEntity" select="getCommentListById"/>
     </resultMap>
 
-    <select id="getCommentListById" parameterType="java.util.Map" resultType="com.capstone.notechigima.domain.CommentDetailEntity">
+    <select id="getCommentListById" parameterType="java.util.Map" resultType="com.capstone.notechigima.domain.comment.CommentDetailEntity">
         SELECT c.comment_id commentId, c.user_id userId,
                u.nickname userName, c.content content,
                c.created_at createdAt, c.updated_at updatedAt
