+++ Notechigima/src/main/resources/mapper/TopicMapper.xml
@@ -11,7 +11,7 @@
         WHERE #{topicId} = t.topic_id
     </select>
 
-    <select id="getTopicList" resultType="com.capstone.notechigima.domain.TopicEntity">
+    <select id="getTopicList" resultType="com.capstone.notechigima.domain.topic.TopicEntity">
         SELECT topic_id topicId, subject_id subjectId, title, updated_at updatedAt, analyzed
         FROM topic s
         WHERE #{subjectId} = s.subject_id
